Version 1.0
Here are the contents of the folders:

Dawson 215B are notes from Dawson's thermal dynamics class 215B.  I took this class in the early 90's

"Dawson Princeton" are notes from John Dawson in the 1960's.  He used it in 1981 as well.

Fried-Nishikawa is a draft of a manuscript that was never published.  The title of the book was "Modern Plasma Physics"

"Morales-222" will be my notes.  I will typeset them in LaTex and incorporate the pictures.

"Banos" is a lecture notes by Alfredo Banos on Asymptotic Methods.